extract item rolling to logic layer - this allows creating a shared frontend formatting function that doesn't need the logic layer as an argument
